+++ QuickConnectPlugin/WinScpDialogSettings.cs
@@ -1,0 +1,11 @@
+namespace QuickConnectPlugin {
+
+    internal class WinScpDialogSettings : AbstractQuickConnectPluginSettings {
+
+        public override void Load() {
+        }
+
+        public override void Save() {
+        }
+    }
+}